refactor(backend): replace Mediator by MediatR and set up FluentValidation
