+++ backend/Zeus.Api.Domain/UserAggregate/ValueObjects/UserId.cs
@@ -1,4 +1,4 @@
-using Zeus.Api.Domain.Common.Models;
+using Zeus.Common.Domain.Models;
 
 namespace Zeus.Api.Domain.UserAggregate.ValueObjects;
