+++ backend/Zeus.Api.Domain/Zeus.Api.Domain.csproj
@@ -8,7 +8,11 @@
 
     <ItemGroup>
       <PackageReference Include="ErrorOr" Version="2.0.1" />
-      <PackageReference Include="Mediator.Abstractions" Version="2.1.7" />
+      <PackageReference Include="MediatR.Contracts" Version="2.0.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Zeus.Common.Domain\Zeus.Common.Domain.csproj" />
     </ItemGroup>
 
 </Project>
