+++ backend/Zeus.Common.Application/Zeus.Common.Application.csproj
@@ -9,12 +9,8 @@
     <ItemGroup>
       <PackageReference Include="ErrorOr" Version="2.0.1" />
       <PackageReference Include="FluentValidation" Version="11.11.0" />
-      <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.11.0" />
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Zeus.Api.Application\Zeus.Api.Application.csproj" />
+      <PackageReference Include="MediatR" Version="12.4.1" />
+      <PackageReference Include="MediatR.Contracts" Version="2.0.1" />
     </ItemGroup>
 
 </Project>
