+++ backend/Zeus.Common.Domain/Models/AggregateRoot.cs
@@ -1,4 +1,4 @@
-namespace Zeus.Api.Domain.Common.Models;
+namespace Zeus.Common.Domain.Models;
 
 public abstract class AggregateRoot<TId>(TId id) : Entity<TId>(id)
     where TId : notnull;
