+++ backend/Zeus.Daemon.Application/DependencyInjection.cs
@@ -1,4 +1,10 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using FluentValidation;
+
+using MediatR;
+
+using Microsoft.Extensions.DependencyInjection;
+
+using Zeus.Common.Application.Behaviors;
 
 namespace Zeus.Daemon.Application;
 
@@ -6,7 +12,9 @@ public static class DependencyInjection
 {
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
-        services.AddMediator();
+        services.AddMediatR(c => c.RegisterServicesFromAssembly(typeof(DependencyInjection).Assembly));
+        services.AddValidatorsFromAssembly(typeof(DependencyInjection).Assembly, ServiceLifetime.Singleton);
+        services.AddSingleton(typeof(IPipelineBehavior<,>), typeof(ValidateBehavior<,>));
         return services;
     }
 }
